restructure again + modify chitchat
